requirements.txt
aiofiles==24.1.0
pyserial-asyncio==0.6
PyYAML==6.0.2
serial==0.0.9